separate input generators in separate file
